@@ -3,8 +3,9 @@ package rscp
 type AuthLevel uint8
 
 // all auth levels as constant
+//
+//nolint:revive,stylecheck
 //go:generate go run github.com/alvaroloes/enumer -type=AuthLevel -json
-//nolint: revive,stylecheck
 const (
 	AUTH_LEVEL_NO_AUTH   AuthLevel = 0
 	AUTH_LEVEL_USER      AuthLevel = 10
