@@ -8,7 +8,8 @@ func (status ClientStatus) String() string {
 }
 
 // all errors as constant
-//nolint: revive,stylecheck
+//
+//nolint:revive,stylecheck
 const (
 	STATUS_OK                        ClientStatus = 0
 	STATUS_ERR_INVALID_INPUT         ClientStatus = -1
