@@ -1,6 +1,6 @@
 package rscp
 
-//nolint: revive,stylecheck
+//nolint:revive,stylecheck
 const (
 	RSCP_FRAME_MAGIC_SIZE  uint16 = 0x2 // magic field size of frame
 	RSCP_FRAME_CTRL_SIZE   uint16 = 0x2 // control field size of frame
@@ -34,18 +34,19 @@ const (
 )
 
 // control field of frame
-//   Byte |        0        |        1        |
-//   Bit  | 7 6 5 4 3 2 1 0 | 7 6 5 4 3 2 1 0 |
-//        | R R R C V V V V | R R R R R R R R |
-//        | R R R           | R R R R R R R R | Reserviert für zukünfige Erweiterungen -> RSCP_CTRL_BIT_MASK
-//        | 0 0 0           | 0 0 0 0 0 0 0 0 | Derzeitiger Zustand für die reservierten Bit
-//        |       C         |                 | Checksummen Flag -> RSCP_CTRL_BIT_MASK
-//        |       0         |                 | Checksumme wird nicht verwendet
-//        |       1         |                 | Checksumme wird verwendet. Das Feld CRC am Ende des Frames ist ein Pflichtfeld, ansonsten wird der Frame verworfen! -> RSCP_CRC_FLAG
-//        |         V V V V |                 | Versionskennzeichnung -> RSCP_CTRL_BIT_MASK_VERSION
-//        |         0 0 0 1 |                 | Version 1.0 (Momentan einzig zugelassener Wert) -> RSCP_VERSION_1_0
 //
-//nolint: lll,revive,stylecheck
+//	Byte |        0        |        1        |
+//	Bit  | 7 6 5 4 3 2 1 0 | 7 6 5 4 3 2 1 0 |
+//	     | R R R C V V V V | R R R R R R R R |
+//	     | R R R           | R R R R R R R R | Reserviert für zukünfige Erweiterungen -> RSCP_CTRL_BIT_MASK
+//	     | 0 0 0           | 0 0 0 0 0 0 0 0 | Derzeitiger Zustand für die reservierten Bit
+//	     |       C         |                 | Checksummen Flag -> RSCP_CTRL_BIT_MASK
+//	     |       0         |                 | Checksumme wird nicht verwendet
+//	     |       1         |                 | Checksumme wird verwendet. Das Feld CRC am Ende des Frames ist ein Pflichtfeld, ansonsten wird der Frame verworfen! -> RSCP_CRC_FLAG
+//	     |         V V V V |                 | Versionskennzeichnung -> RSCP_CTRL_BIT_MASK_VERSION
+//	     |         0 0 0 1 |                 | Version 1.0 (Momentan einzig zugelassener Wert) -> RSCP_VERSION_1_0
+//
+//nolint:lll,revive,stylecheck
 const (
 	RSCP_CTRL_BIT_MASK_CRC     uint16 = 0b0001000000000000 // allowed bit's for crc field of control field
 	RSCP_FLAG_BIT_CRC          uint8  = 12                 // bit start position of crc field of control field
