@@ -1,7 +1,8 @@
 package rscp
 
 // defines the expected Length of the data type or 0 for variable length
-//nolint: gomnd
+//
+//nolint:gomnd
 var lengthMap = map[DataType]uint16{
 	None:      0,
 	Bool:      1,
