@@ -37,7 +37,7 @@ func write(buf *bytes.Buffer, v interface{}) error {
 			// not testable
 			return err
 		}
-		//nolint: gomnd
+		//nolint:gomnd
 		if err := write(buf, int32(v.UnixNano()-(v.Unix()*1e9))); err != nil {
 			// not testable
 			return err
