@@ -127,7 +127,7 @@ func TestWrite(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := Write(&tt.mode, tt.messages, tt.useChecksum)
+			got, err := Write(&tt.mode, tt.messages, tt.useChecksum) // #nosec G601
 			if (err != nil || tt.wantErr != nil) && !(errors.Is(err, tt.wantErr) || err.Error() == tt.wantErr.Error()) {
 				t.Errorf("Write() error = %#v, wantErr %#v", err, tt.wantErr)
 			}
